requirements.txt
openai>=1.0.0
anthropic>=0.7.0
requests>=2.31.0
python-dotenv>=1.0.0
click>=8.1.0
transformers>=4.35.0
torch>=2.0.0
tiktoken>=0.5.0
matplotlib>=3.7.0
tabulate>=0.9.0
colorama>=0.4.6
